requirements.txt
streamlit==1.28.0
pdfminer.six==20221105
python-docx==0.8.11
sentence-transformers==2.2.2
scikit-learn==1.3.0
pandas==2.2.0
numpy==1.26.4

# streamlit==1.28.0
# pdfminer.six==20250327
# python-docx==0.8.11
# sentence-transformers==2.2.2
# scikit-learn==1.3.0
# pandas==2.2.0
# spacy=3.7.5